Clarify project library summaries
Show plain-language Home overview summaries for project libraries while keeping technical locations available on the icon hover. Extract the Home project card row into a focused module with regression coverage for cloud summaries.
