+++ src/registry/contracts/projectLibraries.test.ts
@@ -8,6 +8,8 @@ import {
   getDefaultDirectoryProjectLibraryPath,
   getDefaultDirectoryProjectLibrarySetting,
   getProjectLibraryIdFromSetting,
+  getProjectLibrarySummaryDescription,
+  getProjectLibrarySummaryTooltip,
   LEGACY_PERSONAL_CLOUD_PROJECT_LIBRARY_PATH,
   moveProjectLibrarySetting,
   normalizeProjectLibrarySetting,
@@ -179,6 +181,30 @@ describe('project library settings', () => {
         type: 'directory',
       })
     ).toBe('/projects')
+  })
+
+  test('summarizes project library storage in plain language', () => {
+    const cloudLibrary = {
+      path: DEFAULT_PERSONAL_CLOUD_PROJECT_LIBRARY_LOCAL_PATH,
+      type: 'cloud',
+    }
+    const directoryLibrary = {
+      path: '/projects',
+      type: 'directory',
+    }
+
+    expect(getProjectLibrarySummaryDescription(cloudLibrary)).toBe(
+      'Projects in this library sync to your Zoo account'
+    )
+    expect(getProjectLibrarySummaryTooltip(cloudLibrary)).toContain(
+      'Technical source: zoo://personal'
+    )
+    expect(getProjectLibrarySummaryDescription(directoryLibrary)).toBe(
+      'Projects in this library are saved only on this computer'
+    )
+    expect(getProjectLibrarySummaryTooltip(directoryLibrary)).toContain(
+      'Folder: /projects'
+    )
   })
 
   test('treats the first directory library as the default local project target', () => {
