+++ src/routes/Home.test.tsx
@@ -3,7 +3,9 @@ import {
   FREE_CLOUD_PROJECT_TRAINING_POLICY_URL,
   shouldShowFreeCloudProjectTrainingDisclosure,
 } from '@src/lib/projectLibraries/trainingDisclosure'
+import type { HomeProjectActionsService } from '@src/registry/contracts/homeProjects'
 import { HomeHeader } from '@src/routes/HomeHeader'
+import { ProjectLibraryPreviewRow } from '@src/routes/HomeProjectCards'
 import { render, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 import { describe, expect, it, vi } from 'vitest'
@@ -14,6 +16,22 @@ const cloudLibrary = {
   path: '/documents/Zoo Projects',
   type: 'cloud',
 } satisfies ProjectLibrary
+
+const projectActions = {
+  canOpen: vi.fn(() => false),
+  canDuplicate: vi.fn(() => false),
+  canRename: vi.fn(() => false),
+  canDelete: vi.fn(() => false),
+  canMoveToLibrary: vi.fn(() => false),
+  canReviewDuplicateRealizations: vi.fn(() => false),
+  open: vi.fn(async () => undefined),
+  duplicate: vi.fn(async () => undefined),
+  rename: vi.fn(async () => undefined),
+  delete: vi.fn(async () => undefined),
+  getMoveToLibraryTargets: vi.fn(() => []),
+  moveToLibrary: vi.fn(async () => undefined),
+  deleteDuplicateRealizations: vi.fn(async () => undefined),
+} satisfies HomeProjectActionsService
 
 function renderHomeHeader({
   showFreeCloudProjectTrainingDisclosure,
@@ -32,6 +50,22 @@ function renderHomeHeader({
         showFreeCloudProjectTrainingDisclosure={
           showFreeCloudProjectTrainingDisclosure
         }
+      />
+    </MemoryRouter>
+  )
+}
+
+function renderProjectLibraryPreviewRow(library: ProjectLibrary) {
+  render(
+    <MemoryRouter>
+      <ProjectLibraryPreviewRow
+        library={library}
+        projects={[]}
+        query=""
+        projectStatuses={new Map()}
+        projectActions={projectActions}
+        showCloudSyncUi={true}
+        onMoveToLibrary={vi.fn()}
       />
     </MemoryRouter>
   )
@@ -63,5 +97,21 @@ describe('HomeHeader', () => {
     expect(
       screen.queryByText(/Zoo trains on Free user cloud projects/)
     ).not.toBeInTheDocument()
+  })
+})
+
+describe('ProjectLibraryPreviewRow', () => {
+  it('shows cloud library helper text instead of the technical zoo source', () => {
+    renderProjectLibraryPreviewRow(cloudLibrary)
+
+    const libraryLink = screen.getByTestId('project-library-link')
+    expect(libraryLink).toHaveTextContent('Personal Cloud')
+    expect(libraryLink).toHaveTextContent(
+      'Projects in this library sync to your Zoo account'
+    )
+    expect(libraryLink).not.toHaveTextContent('zoo://personal')
+    expect(
+      screen.getByTitle(/Technical source: zoo:\/\/personal/)
+    ).toBeInTheDocument()
   })
 })
